Fixed circular dependency
