+++ src/apiRoutes/index.ts
@@ -1,3 +1,3 @@
-// export * from './next';
-// export * from './response';
+export * from './next';
+export * from './response';
 export * from './handler';
